Weitere korekturen

+++ tempBerechnung/Berechnung.cs
@@ -5,12 +5,12 @@ namespace tempBerechnung
 {
     public class Berechnung
     {
-        public static double CnachF(int grad)
+        public static double CnachF(double grad)
         {
             return grad * 1.8 + 32;
         }
 
-        public static double FnachC(int grad)
+        public static double FnachC(double grad)
         {
             return (grad - 32)/1.8;
         }
